+++ src/ChilliSource.Cloud.Core/ChilliSource.Cloud.Core.csproj
@@ -27,7 +27,7 @@
 
   <ItemGroup>
     <PackageReference Include="AutoMapper" Version="6.0.2" />
-    <PackageReference Include="ChilliSource.Core.Extensions" Version="1.0.0" />
+    <PackageReference Include="ChilliSource.Core.Extensions" Version="2.0.0" />
     <PackageReference Include="CsvHelper" Version="2.16.3" />
     <PackageReference Include="EnterpriseLibrary.TransientFaultHandling" Version="6.0.1304" />
     <PackageReference Include="EnterpriseLibrary.TransientFaultHandling.Data" Version="6.0.1304.1" />
